Do a temporary fix for the initial ball position
There is a bug when dealing with the ball position after loading a
level. It will initially start at the position here, then moved to the
level specific starting position. The problem is that when the level
specific position is at the same position, the score is counted and the
brick is removed – as it should be. In order to deal with this we move
default position to above the bricks.